requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
prometheus_flask_exporter == 0.23.0
python-logging-loki == 0.3.1
opentelemetry-api == 1.12.0
opentelemetry-sdk == 1.12.0
opentelemetry-instrumentation == 0.33b0
opentelemetry-exporter-jaeger == 1.12.0
opentelemetry-instrumentation-wsgi == 0.33b0
opentelemetry-instrumentation-flask == 0.33b0